@@ -7,31 +7,41 @@
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Citation and citations Collector classes"""
+from __future__ import annotations
 
+from collections import namedtuple
+from functools import wraps
+import logging
 import os
 import sys
-from functools import wraps
-
-from six import iteritems, itervalues
+from typing import TYPE_CHECKING, Any
 
 from .config import DUECREDIT_FILE
 from .entries import DueCreditEntry
-from .stub import InactiveDueCreditCollector
-from .io import TextOutput, PickleOutput
-from .utils import never_fail, borrowdoc
+from .io import PickleOutput, TextOutput
+from .utils import borrowdoc, never_fail
 from .versions import external_versions
-from collections import namedtuple
 
-import logging
-lgr = logging.getLogger('duecredit.collector')
+lgr = logging.getLogger("duecredit.collector")
 
-CitationKey = namedtuple('CitationKey', ['path', 'entry_key'])
+if TYPE_CHECKING:
+    from typing_extensions import Self
 
-class Citation(object):
+CitationKey = namedtuple("CitationKey", ["path", "entry_key"])
+
+
+class Citation:
     """Encapsulates citations and information on their use"""
 
-    def __init__(self, entry, description=None, path=None, version=None,
-                 cite_module=False, tags=['implementation']):
+    def __init__(
+        self,
+        entry: DueCreditEntry,
+        description: str | None = None,
+        path: str | None = None,
+        version: None | str | tuple[str, str] = None,
+        cite_module: bool = False,
+        tags: list[str] | None = None,
+    ):
         """Cite a reference
 
         Parameters
@@ -68,82 +78,84 @@ class Citation(object):
             describing how to contribute some funds to the referenced project
         """
         if path is None:
-            raise ValueError('Must specify path')
+            raise ValueError("Must specify path")
         self._entry = entry
         self._description = description
         # We might want extract all the relevant functionality into a separate class
         self._path = path
         self._cite_module = cite_module
+        if tags is None:
+            tags = ["implementation"]
         self.tags = tags or []
         self.version = version
         self.count = 0
 
-    def __repr__(self):
-        args = [repr(self._entry)]
+    def __repr__(self) -> str:
+        argl = [repr(self._entry)]
         if self._description:
-            args.append("description={0}".format(repr(self._description)))
+            argl.append(f"description={repr(self._description)}")
         if self._path:
-            args.append("path={0}".format(repr(self._path)))
+            argl.append(f"path={repr(self._path)}")
         if self._cite_module:
-            args.append("cite_module={0}".format(repr(self._cite_module)))
+            argl.append(f"cite_module={repr(self._cite_module)}")
 
-        if args:
-            args = ", ".join(args)
+        if argl:
+            args = ", ".join(argl)
         else:
             args = ""
-        return self.__class__.__name__ + '({0})'.format(args)
+        return self.__class__.__name__ + f"({args})"
 
     @property
-    def path(self):
+    def path(self) -> str:
         return self._path
 
-    @property
-    def cite_module(self):
-        return self._cite_module
-
     @path.setter
-    def path(self, path):
+    def path(self, path: str) -> None:
         # TODO: verify value, if we are not up for it -- just make _path public
         self._path = path
 
     @property
-    def entry(self):
+    def entry(self) -> DueCreditEntry:
         return self._entry
 
     @property
-    def description(self):
+    def description(self) -> str | None:
         return self._description
 
     @property
-    def cites_module(self):
-        if not self.path:
-            return None
-        return not (':' in self.path)
+    def cite_module(self) -> bool:
+        return self._cite_module
 
     @property
-    def module(self):
+    def cites_module(self) -> bool | None:
         if not self.path:
             return None
-        return self.path.split(':', 1)[0]
+        return not (":" in self.path)
 
     @property
-    def package(self):
+    def module(self) -> str | None:
+        if not self.path:
+            return None
+        return self.path.split(":", 1)[0]
+
+    @property
+    def package(self) -> str | None:
         module = self.module
         if not module:
             return None
-        return module.split('.', 1)[0]
+        return module.split(".", 1)[0]
 
     @property
-    def objname(self):
+    def objname(self) -> str | None:
         if not self.path:
             return None
-        spl = self.path.split(':', 1)
+        spl = self.path.split(":", 1)
         if len(spl) > 1:
             return spl[1]
         else:
             return None
 
-    def __contains__(self, entry):
+    def __contains__(self, entry: Self) -> bool:
         """Checks if provided entry 'contained' in this one given its path
 
         If current entry is associated with a module, contained will be an entry
@@ -155,26 +167,27 @@ class Citation(object):
         another entry if it really contains it as an attribute
         """
         if self.cites_module:
-            return ((self.path == entry.path) or
-                    (entry.path.startswith(self.path + '.')) or
-                    (entry.path.startswith(self.path + ':')))
+            return (
+                (self.path == entry.path)
+                or (entry.path.startswith(self.path + "."))
+                or (entry.path.startswith(self.path + ":"))
+            )
         else:
-            return entry.path.startswith(self.path + '.')
-
+            return entry.path.startswith(self.path + ".")
 
     @property
-    def key(self):
+    def key(self) -> CitationKey:
         return CitationKey(self.path, self.entry.key)
 
     @staticmethod
-    def get_key(path, entry_key):
+    def get_key(path: str, entry_key: str) -> CitationKey:
         return CitationKey(path, entry_key)
 
-    def set_entry(self, newentry):
+    def set_entry(self, newentry: DueCreditEntry) -> None:
         self._entry = newentry
 
 
-class DueCreditCollector(object):
+class DueCreditCollector:
     """Collect the references
 
     The mighty beast which will might become later a proxy on the way to
@@ -190,12 +203,16 @@ class DueCreditCollector(object):
     """
 
     # TODO?  rename "entries" to "references"?  or "references" is closer to "citations"
-    def __init__(self, entries=None, citations=None):
+    def __init__(
+        self,
+        entries: dict[str, DueCreditEntry] | None = None,
+        citations: dict[CitationKey, Citation] | None = None,
+    ) -> None:
         self._entries = entries or {}
         self.citations = citations or {}
 
     @never_fail
-    def add(self, entry):
+    def add(self, entry: DueCreditEntry | list[DueCreditEntry]) -> None:
         """entry should be a DueCreditEntry object"""
         if isinstance(entry, list):
             for e in entry:
@@ -206,7 +223,7 @@ class DueCreditCollector(object):
             lgr.log(1, "Collector added entry %s", key)
 
     @never_fail
-    def load(self, src):
+    def load(self, src: str) -> None:
         """Loads references from a file or other recognizable source
 
         ATM supported only
@@ -214,14 +231,14 @@ class DueCreditCollector(object):
         """
         # raise NotImplementedError
         if isinstance(src, str):
-            if src.endswith('.bib'):
+            if src.endswith(".bib"):
                 self._load_bib(src)
             else:
-                raise NotImplementedError('Format not yet supported')
+                raise NotImplementedError("Format not yet supported")
         else:
-            raise ValueError('Must be a string')
+            raise ValueError("Must be a string")
 
-    def _load_bib(self, src):
+    def _load_bib(self, src: str) -> None:
         lgr.debug("Loading %s" % src)
 
     # # TODO: figure out what would be the optimal use for the __call__
@@ -234,13 +251,13 @@ class DueCreditCollector(object):
 
     @never_fail
     @borrowdoc(Citation, "__init__")
-    def cite(self, entry, **kwargs):
+    def cite(self, entry: DueCreditEntry | str, **kwargs: Any) -> Citation:
         # TODO: if cite is invoked but no path is provided -- we must figure it out
         # I guess from traceback, otherwise how would we know later to associate it
         # with modules???
-        path = kwargs.get('path', None)
+        path = kwargs.get("path", None)
         if path is None:
-            raise ValueError('path must be provided')
+            raise ValueError("path must be provided")
 
         if isinstance(entry, DueCreditEntry):
             # new one -- add it
@@ -255,26 +272,27 @@ class DueCreditCollector(object):
             citation = self.citations[citation_key]
         except KeyError:
             self.citations[citation_key] = citation = Citation(entry_, **kwargs)
-        assert(citation.key == citation_key)
+        assert isinstance(citation, Citation)
+        assert citation.key == citation_key
         # update citation count
         citation.count += 1
 
         # TODO: theoretically version shouldn't differ if we don't preload previous results
         if not citation.version:
-            version = kwargs.get('version', None)
+            version = kwargs.get("version", None)
 
             if not version and citation.path:
-                modname = citation.path.split('.', 1)[0]
+                modname = citation.path.split(".", 1)[0]
 
-                if '.' in modname:
-                    package = modname.split('.', 1)[0]
+                if "." in modname:
+                    package = modname.split(".", 1)[0]
                 else:
                     package = modname
 
                 # package_loaded = sys.modules.get(package)
                 # if package_loaded:
                 #     # find the citation for that module
-                #     for citation in itervalues(self.citations):
+                #     for citation in self.citations.values():
                 #         if citation.package == package \
                 #                 and not citation.version:
                 version = external_versions[package]
@@ -282,25 +300,25 @@ class DueCreditCollector(object):
 
         return citation
 
-    def _citations_fromentrykey(self):
+    def _citations_fromentrykey(self) -> dict[str, Citation]:
         """Return a dictionary with the current citations indexed by the entry key"""
         citations_key = dict()
-        for (path, entry_key), citation in iteritems(self.citations):
+        for (_path, entry_key), citation in self.citations.items():
             if entry_key not in citations_key:
                 citations_key[entry_key] = citation
 
         return citations_key
 
-
     @staticmethod
-    def _args_match_conditions(conditions, *fargs, **fkwargs):
-        """Helper to identify when to trigger citation given parameters to the function call
-        """
-        for (argpos, kwarg), values in iteritems(conditions):
+    def _args_match_conditions(
+        conditions: dict[Any, Any], *fargs: Any, **fkwargs: Any
+    ) -> bool:
+        """Helper to identify when to trigger citation given parameters to the function call"""
+        for (argpos, kwarg), values in conditions.items():
             # main logic -- assess default and if get to the next one if
             # given argument is not present
             if not ((len(fargs) > argpos) or (kwarg in fkwargs)):
-                if not ('DC_DEFAULT' in values):
+                if not ("DC_DEFAULT" in values):
                     # if value was specified but not provided and not default
                     # conditions are not satisfied
                     return False
@@ -312,12 +330,12 @@ class DueCreditCollector(object):
                 value = fargs[argpos]
             if kwarg in fkwargs:
                 value = fkwargs[kwarg]
-            assert(value != "__duecredit_magical_undefined__")
+            assert value != "__duecredit_magical_undefined__"
 
-            if '.' in kwarg:
+            if "." in kwarg:
                 # we were requested to condition based on the value of the attribute
                 # of the value.  So get to the attribute(s) value
-                for attr in kwarg.split('.')[1:]:
+                for attr in kwarg.split(".")[1:]:
                     value = getattr(value, attr)
 
             # Value is present but not matching
@@ -369,7 +387,7 @@ class DueCreditCollector(object):
 
         Conditional based on the state of the object
 
-        >>> class Citeable(object):
+        >>> class Citeable:
         ...     def __init__(self, param=None):
         ...         self.param = param
         ...     @due.dcite('XXX00', description="The same good old relief",
@@ -379,28 +397,31 @@ class DueCreditCollector(object):
         >>> Citeable('magic')([1, 2])
         3
         """
+
         def func_wrapper(func):
-            conditions = kwargs.pop('conditions', {})
-            path = kwargs.get('path', None)
+            conditions = kwargs.pop("conditions", {})
+            path = kwargs.get("path", None)
             if not path:
                 # deduce path from the actual function which was decorated
                 # TODO: must include class name  but can't !!!???
                 modname = func.__module__
-                path = kwargs['path'] = '%s:%s' % (modname, func.__name__)
+                path = kwargs["path"] = "{}:{}".format(modname, func.__name__)
             else:
                 # TODO: we indeed need to separate path logic outside
-                modname = path.split(':', 1)[0]
+                modname = path.split(":", 1)[0]
 
             # if decorated function was invoked, it means that we need
             # to cite that even if it is a module. But do not override
             # value if user explicitly stated
-            if 'cite_module' not in kwargs:
-                kwargs['cite_module'] = True
+            if "cite_module" not in kwargs:
+                kwargs["cite_module"] = True
 
             # TODO: might make use of inspect.getmro
             # see e.g.
             # http://stackoverflow.com/questions/961048/get-class-that-defined-method
-            lgr.debug("Decorating func %s within module %s" % (func.__name__, modname))
+            lgr.debug(
+                "Decorating func {} within module {}".format(func.__name__, modname)
+            )
             # TODO: unittest for all the __version__ madness
 
             # TODO: check if we better use wrapt module which provides superior "correctness"
@@ -408,55 +429,63 @@ class DueCreditCollector(object):
             @wraps(func)
             def cite_wrapper(*fargs, **fkwargs):
                 try:
-                    if not conditions \
-                            or self._args_match_conditions(conditions, *fargs, **fkwargs):
-                        citation = self.cite(*args, **kwargs)
+                    if not conditions or self._args_match_conditions(
+                        conditions, *fargs, **fkwargs
+                    ):
+                        self.cite(*args, **kwargs)
                 except Exception as e:
-                    lgr.warning("Failed to cite due to %s" % (e,))
+                    lgr.warning("Failed to cite due to {}".format(e))
                 return func(*fargs, **fkwargs)
 
             cite_wrapper.__duecredited__ = func
             return cite_wrapper
+
         return func_wrapper
 
     @never_fail
-    def __repr__(self):
-        args = []
+    def __repr__(self) -> str:
+        argl = []
         if self.citations:
-            args.append("citations={0}".format(repr(self.citations)))
+            argl.append(f"citations={repr(self.citations)}")
         if self._entries:
-            args.append("entries={0}".format(repr(self._entries)))
+            argl.append(f"entries={repr(self._entries)}")
 
-        if args:
-            args = ", ".join(args)
+        if argl:
+            args = ", ".join(argl)
         else:
             args = ""
-        return self.__class__.__name__ + '({0})'.format(args)
+        return self.__class__.__name__ + f"({args})"
 
     @never_fail
-    def __str__(self):
-        return self.__class__.__name__ + \
-            ' {0:d} entries, {1:d} citations'.format(
-                len(self._entries), len(self.citations))
+    def __str__(self) -> str:
+        return self.__class__.__name__ + " {:d} entries, {:d} citations".format(
+            len(self._entries), len(self.citations)
+        )
 
 
 # TODO: redo heavily -- got very messy
-class CollectorSummary(object):
-    """A helper which would take care about exporting citations upon its Death
-    """
-    def __init__(self, collector, outputs="stdout,pickle", fn=DUECREDIT_FILE):
+class CollectorSummary:
+    """A helper which would take care about exporting citations upon its Death"""
+
+    def __init__(
+        self,
+        collector: DueCreditCollector,
+        outputs: str = "stdout,pickle",
+        fn: str = DUECREDIT_FILE,
+    ) -> None:
         self._due = collector
         self.fn = fn
         # for now decide on output "format" right here
         self._outputs = [
-            self._get_output_handler(
-                type_.lower().strip(), collector, fn=fn)
-            for type_ in os.environ.get('DUECREDIT_OUTPUTS', outputs).split(',')
+            self._get_output_handler(type_.lower().strip(), collector, fn=fn)
+            for type_ in os.environ.get("DUECREDIT_OUTPUTS", outputs).split(",")
             if type_
         ]
 
     @staticmethod
-    def _get_output_handler(type_, collector, fn=None):
+    def _get_output_handler(
+        type_: str, collector: DueCreditCollector, fn: str | None = None
+    ) -> TextOutput | PickleOutput:
         # just a little factory
         if type_ in ("stdout", "stderr"):
             return TextOutput(getattr(sys, type_), collector)
@@ -465,10 +494,9 @@ class CollectorSummary(object):
         else:
             raise NotImplementedError()
 
-    def dump(self):
+    def dump(self) -> None:
         for output in self._outputs:
             output.dump()
 
 
-# TODO:  provide HTML, MD, RST etc formattings
-
+# TODO:  provide HTML, MD, RST etc output formats
