@@ -7,88 +7,88 @@
 #   under MIT license
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
-import re
-from six import PY2
+from __future__ import annotations
 
 import logging
-lgr = logging.getLogger('duecredit.entries')
+import re
+
+lgr = logging.getLogger("duecredit.entries")
 
 
-class DueCreditEntry(object):
-    def __init__(self, rawentry, key=None):
+class DueCreditEntry:
+    _key: str
+
+    def __init__(self, rawentry: str, key: str | None = None) -> None:
         self._rawentry = rawentry
         self._key = key or rawentry.lower()
 
-    def __eq__(self, other):
-        return (
-            (self._rawentry == other._rawentry) and
-            (self._key == other._key)
-        )
+    def __eq__(self, other: object) -> bool:
+        if not isinstance(other, DueCreditEntry):
+            return NotImplemented
+        return (self._rawentry == other._rawentry) and (self._key == other._key)
 
-    def get_key(self):
+    def get_key(self) -> str:
         return self._key
 
     @property
-    def key(self):
+    def key(self) -> str:
         return self.get_key()
 
     @property
-    def rawentry(self):
-        if PY2:
-            return unicode(self._rawentry)
-        else:
-            return self._rawentry
+    def rawentry(self) -> str:
+        return self._rawentry
 
-    def _process_rawentry(self):
+    def _process_rawentry(self) -> None:
         pass
 
-    def __repr__(self):
-        args = [repr(self._rawentry),
-                "key={0}".format(repr(self._key))]
-        args = ", ".join(args)
-        return self.__class__.__name__ + '({0})'.format(args)
+    def __repr__(self) -> str:
+        argl = [repr(self._rawentry), f"key={repr(self._key)}"]
+        args = ", ".join(argl)
+        return self.__class__.__name__ + f"({args})"
 
-    def format(self):
+    def format(self) -> str:
         # TODO: return nice formatting of the entry
         return str(self._rawentry)
 
 
 class BibTeX(DueCreditEntry):
-    def __init__(self, bibtex, key=None):
-        super(BibTeX, self).__init__(bibtex.strip())
-        self._key = None
+    def __init__(self, bibtex: str, key: str | None = None) -> None:
+        super().__init__(bibtex.strip())
         self._reference = None
         self._process_rawentry()
         if key is not None:
             # use the one provided, not the parsed one
-            lgr.debug("Replacing parsed key %s for BibTeX with the provided %s",
-                      self._key, key)
+            lgr.debug(
+                "Replacing parsed key %s for BibTeX with the provided %s",
+                self._key,
+                key,
+            )
             self._key = key
 
-    def _process_rawentry(self):
-        reg = re.match(r"\s*@(?P<type>\S*)\s*\{\s*(?P<key>\S*)\s*,.*",
-                       self._rawentry, flags=re.MULTILINE)
-        assert(reg)
+    def _process_rawentry(self) -> None:
+        reg = re.match(
+            r"\s*@(?P<type>\S*)\s*\{\s*(?P<key>\S*)\s*,.*",
+            self._rawentry,
+            flags=re.MULTILINE,
+        )
+        assert reg
         matches = reg.groupdict()
-        self._key = matches['key']
+        self._key = matches["key"]
 
 
 class Text(DueCreditEntry):
-    """Just a free text entry without any special super powers in rendering etc
-    """
+    """Just a free text entry without any special super powers in rendering etc"""
+
     pass  # nothing special I guess
 
 
 class Doi(DueCreditEntry):
-
     @property
-    def doi(self):
+    def doi(self) -> str:
         return self._rawentry
 
 
 class Url(DueCreditEntry):
-
     @property
-    def url(self):
+    def url(self) -> str:
         return self._rawentry
-
