@@ -1,11 +1,12 @@
 """Some test submodule"""
+from typing import Any
 
 
-def testfunc(arg1, kwarg1=None):
+def testfunc(arg1: Any, kwarg1: Any = None) -> str:
     """testfunc docstring"""
-    return "testfunc: %s, %s" % (arg1, kwarg1)
+    return "testfunc: {}, {}".format(arg1, kwarg1)
 
 
-class TestClass(object):
-    def testmeth(self, arg1, kwarg1=None):
-        return "testmeth: %s, %s" % (arg1, kwarg1)
+class TestClass:
+    def testmeth(self, arg1: Any, kwarg1: Any = None) -> str:
+        return "testmeth: {}, {}".format(arg1, kwarg1)
