@@ -9,44 +9,56 @@
 """Spit out the summary of the citations which were collected.
 
 """
+from __future__ import annotations
 
-import sys
+import argparse
 import os
+import sys
 
-from ..log import lgr
 from ..config import DUECREDIT_FILE
-from ..collector import CollectorSummary
-from ..io import TextOutput, BibTeXOutput
+from ..io import BibTeXOutput, TextOutput
+from ..log import lgr
 
-__docformat__ = 'restructuredtext'
+__docformat__ = "restructuredtext"
 
 # magic line for manpage summary
 # man: -*- % summary of collected citations
 
-def setup_parser(parser):
+
+def setup_parser(parser: argparse.ArgumentParser) -> None:
+    parser.add_argument(
+        "-f",
+        "--filename",
+        default=DUECREDIT_FILE,
+        help="Filename containing collected citations. Default: %(default)s",
+    )
 
     parser.add_argument(
-        "-f", "--filename", default=DUECREDIT_FILE,
-        help="Filename containing collected citations. Default: %(default)s")
+        "--style",
+        choices=("apa", "harvard1"),
+        default="harvard1",
+        help="Style to be used for listing citations",
+    )
 
     parser.add_argument(
-        "--style", choices=("apa", "harvard1"), default="harvard1",
-        help="Style to be used for listing citations")
+        "--format",
+        choices=("text", "bibtex"),
+        default="text",
+        help="Way to present the summary",
+    )
 
-    parser.add_argument(
-        "--format", choices=("text", "bibtex"), default="text",
-        help="Way to present the summary")
 
-def run(args):
+def run(args: argparse.Namespace) -> int:
     from ..io import PickleOutput
+
     if not os.path.exists(args.filename):
-        lgr.debug("File {0} doesn't exist.  No summary available".format(
-            args.filename))
+        lgr.debug("File {} doesn't exist.  No summary available".format(args.filename))
         return 1
 
     due = PickleOutput.load(args.filename)
-    #CollectorSummary(due).dump()
+    # CollectorSummary(due).dump()
 
+    out: TextOutput | BibTeXOutput
     if args.format == "text":
         out = TextOutput(sys.stdout, due, args.style)
     elif args.format == "bibtex":
@@ -54,5 +66,4 @@ def run(args):
     else:
         raise ValueError("unknown to treat %s" % args.format)
     out.dump()
-
-
+    return 0
