@@ -7,22 +7,23 @@
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Run internal DueCredit (unit)tests to verify correct operation on the system"""
+from __future__ import annotations
 
+__docformat__ = "restructuredtext"
 
-__docformat__ = 'restructuredtext'
+import argparse
 
 # magic line for manpage summary
 # man: -*- % run unit-tests
 
-from .helpers import parser_add_common_args
 
-
-def setup_parser(parser):
+def setup_parser(parser: argparse.ArgumentParser) -> None:
     # TODO -- pass options such as verbosity etc
     pass
 
 
-def run(args):
-    import duecredit
+def run(_args: argparse.Namespace) -> None:
+    import duecredit  # noqa: F401
+
     raise NotImplementedError("Just use pytest duecredit for now")
-    #duecredit.test()
+    # duecredit.test()
