@@ -9,7 +9,8 @@
 """Facility to automagically decorate with references other modules
 """
 
-__docformat__ = 'restructuredtext'
+__docformat__ = "restructuredtext"
 
 from .injector import DueCreditInjector
 
+__all__ = ["DueCreditInjector"]
