@@ -9,19 +9,27 @@
 """
 Automatic injection of bibliography entries for dipy module
 """
+from __future__ import annotations
 
+from typing import TYPE_CHECKING
 
-from ..entries import Doi, BibTeX
+from ..entries import Doi
 
 # If defined, would determine from which to which version of the corresponding
 # module to care about
 min_version = None
 max_version = None
 
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
-def inject(injector):
-    #http://nipy.org/dipy/cite.html#a-note-on-citing-our-work
-    injector.add('dipy', None, Doi('10.3389/fninf.2014.00008'),
-                    description='Dipy, a library for the analysis of diffusion MRI data.',
-                    tags=['implementation'])
 
+def inject(injector: DueCreditInjector) -> None:
+    # http://nipy.org/dipy/cite.html#a-note-on-citing-our-work
+    injector.add(
+        "dipy",
+        None,
+        Doi("10.3389/fninf.2014.00008"),
+        description="Dipy, a library for the analysis of diffusion MRI data.",
+        tags=["implementation"],
+    )
