@@ -9,14 +9,19 @@
 """
 Automatic injection of bibliography entries for matplotlib module
 """
+from __future__ import annotations
 
-from ..entries import Doi, BibTeX
+from typing import TYPE_CHECKING
 
+from ..entries import BibTeX, Doi
 
 # If defined, would determine from which to which version of the corresponding
 # module to care about
 min_version = None
 max_version = None
+
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
 bib_str = """
 @Article{Hunter:2007,
@@ -36,14 +41,22 @@ bib_str = """
 """.strip()
 
 
-def inject(injector):
-    injector.add("matplotlib", None, BibTeX(bib_str), description="Plotting with Python", tags=["implementation"])
+def inject(injector: DueCreditInjector) -> None:
+    injector.add(
+        "matplotlib",
+        None,
+        BibTeX(bib_str),
+        description="Plotting with Python",
+        tags=["implementation"],
+    )
 
     doi_prefix = "10.5281/zenodo."
 
     # latest version
     injector.add(
-        "matplotlib", None, Doi(doi_prefix + "2893252"),
-        description="Plotting with Python", tags=["implementation"]
+        "matplotlib",
+        None,
+        Doi(doi_prefix + "2893252"),
+        description="Plotting with Python",
+        tags=["implementation"],
     )
-
