@@ -9,21 +9,34 @@
 """
 Automatic injection of bibliography entries for mne module
 """
+from __future__ import annotations
 
+from typing import TYPE_CHECKING
 
-from ..entries import Doi, BibTeX
+from ..entries import Doi
 
 # If defined, would determine from which to which version of the corresponding
 # module to care about
 min_version = None
 max_version = None
 
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
-def inject(injector):
-    #http://martinos.org/mne/stable/cite.html
-    injector.add('mne', None, Doi('10.1016/j.neuroimage.2013.10.027'),
-                    description='MNE software for processing MEG and EEG data.',
-                    tags=['implementation'])
-    injector.add('mne', None, Doi('10.3389/fnins.2013.00267'),
-                    description='MEG and EEG data analysis with MNE-Python.',
-                    tags=['implementation'])
+
+def inject(injector: DueCreditInjector) -> None:
+    # http://martinos.org/mne/stable/cite.html
+    injector.add(
+        "mne",
+        None,
+        Doi("10.1016/j.neuroimage.2013.10.027"),
+        description="MNE software for processing MEG and EEG data.",
+        tags=["implementation"],
+    )
+    injector.add(
+        "mne",
+        None,
+        Doi("10.3389/fnins.2013.00267"),
+        description="MEG and EEG data analysis with MNE-Python.",
+        tags=["implementation"],
+    )
