@@ -8,7 +8,9 @@
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Automatic injection of bibliography entries for nibabel module
 """
+from __future__ import annotations
 
+from typing import TYPE_CHECKING
 
 from ..entries import Doi
 
@@ -17,10 +19,16 @@ from ..entries import Doi
 min_version = None
 max_version = None
 
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
-def inject(injector):
-    injector.add('nibabel', None,
-                 Doi('10.5281/zenodo.60847'),
-                 cite_module=True,
-                 description="I/O library to access to common neuroimaging file formats",
-                 tags=['implementation'])
+
+def inject(injector: DueCreditInjector) -> None:
+    injector.add(
+        "nibabel",
+        None,
+        Doi("10.5281/zenodo.60847"),
+        cite_module=True,
+        description="I/O library to access to common neuroimaging file formats",
+        tags=["implementation"],
+    )
