@@ -8,17 +8,27 @@
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 """Automatic injection of bibliography entries for pandas module
 """
+from __future__ import annotations
 
-from ..entries import Doi, BibTeX, Url
+from typing import TYPE_CHECKING
+
+from ..entries import BibTeX
 
 # If defined, would determine from which to which version of the corresponding
 # module to care about
 min_version = None
 max_version = None
 
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
-def inject(injector):
-    injector.add('pandas', None, BibTeX("""
+
+def inject(injector: DueCreditInjector) -> None:
+    injector.add(
+        "pandas",
+        None,
+        BibTeX(
+            """
         @InProceedings{ mckinney-proc-scipy-2010,
           author    = { McKinney, Wes },
           title     = { Data Structures for Statistical Computing in Python },
@@ -27,4 +37,7 @@ def inject(injector):
           year      = { 2010 },
           editor    = { van der Walt, St\'efan and Millman, Jarrod }
         }
-    """), description="Data analysis library for tabular data")
+    """
+        ),
+        description="Data analysis library for tabular data",
+    )
