@@ -9,18 +9,27 @@
 """
 Automatic injection of bibliography entries for skimage module
 """
+from __future__ import annotations
 
+from typing import TYPE_CHECKING
 
-from ..entries import Doi, BibTeX
+from ..entries import Doi
 
 # If defined, would determine from which to which version of the corresponding
 # module to care about
 min_version = None
 max_version = None
 
+if TYPE_CHECKING:
+    from .injector import DueCreditInjector
 
-def inject(injector):
-    #http://scikit-image.org
-    injector.add('skimage', None, Doi('10.7717/peerj.453'),
-                    description='scikit-image: Image processing in Python.',
-                    tags=['implementation'])
+
+def inject(injector: DueCreditInjector) -> None:
+    # http://scikit-image.org
+    injector.add(
+        "skimage",
+        None,
+        Doi("10.7717/peerj.453"),
+        description="scikit-image: Image processing in Python.",
+        tags=["implementation"],
+    )
