@@ -1,1 +1,1 @@
-raise ImportError
+raise ImportError("Artificially failining import as if no lxml is available")
