@@ -1,5 +1,5 @@
 """Module to test various duecredit functionality, e.g. injections"""
 
-from .imported import *
+from .imported import *  # noqa: F401, F403
 
-__version__ = '0.5'
+__version__ = "0.5"
