@@ -10,7 +10,7 @@ from sklearn.datasets import make_blobs
 print("I: Simulating 4 blobs")
 data, true_label = make_blobs(centers=4)
 
-dist = pdist(data, metric='euclidean')
+dist = pdist(data, metric="euclidean")
 
-Z = linkage(dist, method='single')
+Z = linkage(dist, method="single")
 print("I: Done clustering 4 blobs")
